Update README.md to add grempe/niceware-ts TypeScript port
Hi, I've forked `Niceware` to `Niceware TS` to port it to TypeScript and add a number of modernizations that make it type safe and improve the developer experience. It no longer has dependencies on Node API's. It remains fully compatible and now has a modern test suite that has 100% code coverage (including the wordlist) and benchmark tests. It now exports CommonJS and ESM modules so it can easily be used in modern tooling including Deno.

See the README here: https://github.com/grempe/niceware-ts